requirements.txt
autopep8==1.4.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2018.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
stripe==2.51.0
urllib3==1.25.10
whitenoise==5.2.0